Day 17/Plan_the_trip/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.28.1
requests==2.32.3
langgraph==0.2.60
langchain-core==0.3.29
langchain-openai==0.2.14
langsmith==0.2.10
reportlab==4.2.5
faiss-cpu==1.9.0.post1
numpy==2.2.1
